Added module pkcs11
The example project com.itextpdf.signingexamples:signing-examples-pkcs11 contains example code for signing PDFs with iText 7 using a PKCS#11 device with the SunPKCS11 provider (and in some cases the IAIK PKCS#11 provider or the IAIK PKCS#11 wrapper) installed. It has been created somewhat in parallel to the com.itextpdf.signingexamples:signing-examples-pkcs11-java11 project (in module pkcs11-java11); the Java 11 project actually is the lead project of these two.

Beware, this module has numerous external dependencies; on one hand there are the IAIK libraries which cannot all be retrieved from Maven Central due to their licenses; on the other hand there are a number of PKCS#11 modules provided with the respective device. Please read the README.md file for details.