Changed assumption on CryptoServerJCE.jar location
The CryptoServerJCE.jar Utimaco security provider jar had been assumed to be located in lib subfolder of the jce-utimaco module. This has been changed: Now it is assumed to be resolvable by Maven, either locally or from a remote repository. Details can be found in the README.md file of the module.